Update to Java 17, MC 1.18, and bump version to 2.0
The primary change in this commit is to re-do the parts of the plugin
that interacted with the org.ajave ebean dependency. This has now been
replaced with the successor project io.ebeans. No major changes
otherwise, beside the fact that initial installation requires manual
work. This can be improved later perhaps, but this feature was removed
from the underlying library, and the intent of this release is to do
minimal changes. Due to this, however, the version has been bumped to
2.0 as this is a somewhat "breaking" change. On first run, the necessary
SQL is printed to the logs, which must be run manually.